Enhance security and update user interface
This commit adds measures to prevent direct access to several PHP pages, enhancing the application's security. On the user interface side, it includes changing 'Filler' to 'Instruments' in the navigation menu and adding a footer linking to the project's GitHub page. Modifications on CSS layout properties are also done to improve the page design and responsiveness.